+++ packages/core/rsc/index.ts
@@ -6,6 +6,7 @@ export type {
   render,
   createAI,
 } from './rsc-server';
+
 export type {
   readStreamableValue,
   useUIState,
@@ -13,3 +14,5 @@ export type {
   useActions,
   useSyncUIState,
 } from './rsc-client';
+
+export type { StreamableValue } from './types';